Read slices in DU

+++ openair2/COMMON/f1ap_messages_types.h
@@ -132,7 +132,7 @@ typedef struct f1ap_served_cell_info_t {
   // Number of slide support items (max 16, could be increased to as much as 1024)
   uint16_t num_ssi;
   uint8_t sst;
-  uint8_t sd;
+  uint32_t sd;
 
   f1ap_mode_t mode;
   union {
